data/requirements.txt
Pillow
BeautifulSoup
scipy
requests